included appropriation for  covid19 data from JHU

+++ src/Covid19Reports.Lib/Publisher/ComparisonReportPublisher.cs
@@ -93,7 +93,17 @@ namespace Covid19Reports.Lib.Publisher
 
             template = template.Replace("RECOVERYPROGRESSBARWIDTH",recoveryProgressBarWidth.ToString());
 
-            File.WriteAllText(reportName,template);
+             //Update or overwrite a file only if has changed
+            if (!File.Exists(reportName))
+                File.WriteAllText(reportName,template);
+            else
+            {
+                var existingFileContent = File.ReadAllText(reportName);
+
+                if (!existingFileContent.Equals(template))
+                     File.WriteAllText(reportName,template);
+            }
+   
         }
     }
 }
